finish extract account model with unittest
