+++ omise/__init__.py
@@ -25,7 +25,6 @@ api_vault = 'https://vault.omise.co'
 
 
 __all__ = [
-    'Account',
     'Balance',
     'BankAccount',
     'Card',
@@ -213,43 +212,6 @@ class _VaultResource(Base):
     @classmethod
     def _request(cls, *args, **kwargs):
         return Request(api_public, api_vault, api_version).send(*args, **kwargs)
-
-
-class Account(_MainResource, Base):
-    """API class representing accounts details.
-
-    This API class is used for retrieving account information such as creator
-    email or account creation date. The account retrieved by this API is the
-    account associated with API secret key.
-
-    Basic usage::
-
-        >>> import omise
-        >>> omise.api_secret = 'skey_test_4xs8breq3htbkj03d2x'
-        >>> account = omise.Account.retrieve()
-        <Account id='acct_4xs8bre8a8vhrgijcjg' at 0x7f7410021990>
-        >>> account.email
-        None
-    """
-
-    @classmethod
-    def _instance_path(cls, *args):
-        return 'account'
-
-    @classmethod
-    def retrieve(cls):
-        """Retrieve the account details associated with the API key.
-
-        :rtype: Account
-        """
-        return _as_object(cls._request('get', cls._instance_path()))
-
-    def reload(self):
-        """Reload the account details.
-
-        :rtype: Account
-        """
-        return self._reload_data(self._request('get', self._instance_path()))
 
 
 class Balance(_MainResource, Base):
@@ -1890,3 +1852,6 @@ class Transaction(_MainResource, Base):
         return self._reload_data(
             self._request('get',
                           self._instance_path(self._attributes['id'])))
+
+
+from .resources.account import Account
